% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/r_credit_card_numbers.R
\name{r_credit_card_numbers}
\alias{r_credit_card_numbers}
\title{Generate random fake credit card numbers.}
\usage{
r_credit_card_numbers(n)
}
\arguments{
\item{n}{number of observations.}
}
\value{
A character vector of \code{n} randomly generated credit card numbers.
}
\description{
Generate random fake credit card numbers.
}
\examples{
r_credit_card_numbers(10)
}

