% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/r_date_of_births.R
\name{r_date_of_births}
\alias{r_date_of_births}
\title{Generate random fake date of birth values.}
\usage{
r_date_of_births(n, start = as.Date("1900-01-01"), end = Sys.Date())
}
\arguments{
\item{n}{number of observations.}

\item{start}{starting date.}

\item{end}{ending date.}
}
\value{
A character vector of \code{n} randomly generated date of birth values.
}
\description{
Generate random fake date of birth values.
}
\examples{
r_date_of_births(10)
r_date_of_births(10, start = as.Date("2000-01-01"))
r_date_of_births(10,
                 start = as.Date("2000-01-01"),
                 end = as.Date("2100-01-01"))
}

