% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/r_email_addresses.R
\name{r_email_addresses}
\alias{r_email_addresses}
\title{Generate random fake e-mail addresses.}
\usage{
r_email_addresses(n)
}
\arguments{
\item{n}{number of observations.}
}
\value{
A character vector of \code{n} fake randomly generated e-mail addresses.
}
\description{
Generate random fake e-mail addresses.
}

