% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/r_full_names.R
\name{r_full_names}
\alias{r_full_names}
\title{Generate random fake full names.}
\usage{
r_full_names(n)
}
\arguments{
\item{n}{number of observations.}
}
\value{
A character vector of \code{n} fake randomly generated full names.
}
\description{
Generate random fake full names.
}

