% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/r_ip_addresses.R
\name{r_ipv4_addresses}
\alias{r_ipv4_addresses}
\title{Generate random fake IPv4 address numbers.}
\usage{
r_ipv4_addresses(n)
}
\arguments{
\item{n}{number of observations.}
}
\value{
A character vector of \code{n} randomly generated IP address numbers.
}
\description{
Generate random fake IPv4 address numbers.
}

