% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/r_latitudes.R
\name{r_latitudes}
\alias{r_latitudes}
\title{Generate random fake latitude values.}
\usage{
r_latitudes(n)
}
\arguments{
\item{n}{number of observations.}
}
\value{
A character vector of \code{n} randomly generated latitude values.
}
\description{
Generate random fake latitude values.
}

