% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/r_longitudes.R
\name{r_longitudes}
\alias{r_longitudes}
\title{Generate random fake longitude values.}
\usage{
r_longitudes(n)
}
\arguments{
\item{n}{number of observations.}
}
\value{
A character vector of \code{n} randomly generated longitude values.
}
\description{
Generate random fake longitude values.
}
\examples{
r_longitudes(10)
}

