% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/r_national_identification_numbers.R
\name{r_national_identification_numbers}
\alias{r_national_identification_numbers}
\title{Generate random fake national identification numbers.}
\usage{
r_national_identification_numbers(n)
}
\arguments{
\item{n}{number of observations.}
}
\value{
A character vector of \code{n} randomly generated national identification numbers.
}
\description{
Generate random fake national identification numbers.
}
\examples{
r_national_identification_numbers(10)
}

