% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/r_phone_numbers.R
\name{r_phone_numbers}
\alias{r_phone_numbers}
\title{Generate random fake phone numbers.}
\usage{
r_phone_numbers(n, use_hyphens = FALSE, use_parentheses = FALSE,
  use_spaces = FALSE)
}
\arguments{
\item{n}{number of observations.}

\item{use_hyphens}{should hyphens be included.}

\item{use_parentheses}{should parantheses be included.}

\item{use_spaces}{should spaces be included.}
}
\value{
A character vector of \code{n} randomly generated phone numbers.
}
\description{
Generate random fake phone numbers.
}

