% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/r_vehicle_registration_plate_numbers.R
\name{r_vehicle_registration_plate_numbers}
\alias{r_vehicle_registration_plate_numbers}
\title{Generate random fake vehicle registration plate numbers.}
\usage{
r_vehicle_registration_plate_numbers(n)
}
\arguments{
\item{n}{number of observations.}
}
\value{
A character vector of \code{n} randomly generated vehicle registration plate numbers.
}
\description{
Generate random fake vehicle registration plate numbers.
}

